removed only raw and hdr files

+++ src/test/removefiles.py
@@ -1,10 +1,17 @@
 import os
 import time
 
-path = '/mnt/ramdisk'
+path = '/mnt/ramdisk/'
 
 while True:
-    for f in os.listdir(path):
-        print "Removing: "+str(f)
-        os.remove(os.path.join(path, f))
-    time.sleep(0.1)
+    allfiles = sorted(os.listdir(path))
+    if len(allfiles) > 20:
+        for f in allfiles[:10]:
+            if f.endswith('raw'):
+                print "Removing: "+str(f)
+                os.remove(os.path.join(path, f))
+            elif f.endswith('hdr'):
+                print "Removing: "+str(f)
+                os.remove(os.path.join(path, f))
+        time.sleep(1)
+    time.sleep(1)
